Sharing sbe-rest code with contact app.

+++ app/contact-app/contact-rest-services/src/main/java/org/springbyexample/contact/web/client/AbstractPersistenceFindClient.java
@@ -21,32 +21,32 @@ import java.util.Map;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-import org.springbyexample.contact.web.service.PersistenceFindMarshallingService;
+import org.springbyexample.mvc.rest.service.PersistenceFindMarshallingService;
 import org.springbyexample.schema.beans.response.EntityFindResponseResult;
 import org.springbyexample.schema.beans.response.EntityResponseResult;
 
 
 /**
  * Abstract persistence find client.
- * 
+ *
  * @author David Winterfeldt
- * 
+ *
  *  @param   <R>      Generic response.
  *  @param   <FR>     Find response.
  */
-public abstract class AbstractPersistenceFindClient<R extends EntityResponseResult, FR extends EntityFindResponseResult> 
+public abstract class AbstractPersistenceFindClient<R extends EntityResponseResult, FR extends EntityFindResponseResult>
         implements PersistenceFindMarshallingService<R, FR> {
 
     final Logger logger = LoggerFactory.getLogger(getClass());
-    
+
     protected final RestClient client;
-    
+
     private final String findByIdRequest;
     private final String findPaginatedRequest;
     private final String findRequest;
     protected final Class<R> responseClazz;
     protected final Class<FR> findResponseClazz;
-    
+
 
     public AbstractPersistenceFindClient(RestClient client,
                                          String findByIdRequest, String findPaginatedRequest, String findRequest,
@@ -60,42 +60,42 @@ public abstract class AbstractPersistenceFindClient<R extends EntityResponseResu
     }
 
     @Override
-    public R findById(long id) {
+    public R findById(Integer id) {
         R response = null;
-        
+
         String url = client.createUrl(findByIdRequest);
-        
+
         logger.debug("REST client findById.  id={}  url='{}'", id, url);
-        
+
         response = client.getRestTemplate().getForObject(url, responseClazz, createPkVars(id));
-        
+
         return response;
     }
 
     @Override
     public FR find(int page, int pageSize) {
         FR response = null;
-        
+
         String url = client.createUrl(findPaginatedRequest);
-        
-        logger.debug("REST client paginated find.  page={}  pageSize={}  url='{}'", 
+
+        logger.debug("REST client paginated find.  page={}  pageSize={}  url='{}'",
                      new Object[] { page, pageSize, url});
 
         response = client.getRestTemplate().getForObject(url, findResponseClazz, createPageVars(page, pageSize));
-        
+
         return response;
     }
 
     @Override
     public FR find() {
         FR response = null;
-        
+
         String url = client.createUrl(findRequest);
-        
+
         logger.debug("REST client find. url='{}'", url);
 
         response = client.getRestTemplate().getForObject(url, findResponseClazz);
-        
+
         return response;
     }
 
@@ -111,10 +111,10 @@ public abstract class AbstractPersistenceFindClient<R extends EntityResponseResu
      */
     public Map<String, Integer> createPageVars(int page, int pageSize) {
         Map<String, Integer> result = new HashMap<String, Integer>();
-        
+
         result.put(PAGE_VAR, page);
         result.put(PAGE_SIZE_VAR, pageSize);
-        
+
         return result;
     }
